+++ crates/console/src/lib.rs
@@ -5,6 +5,7 @@ mod console;
 #[cfg(feature = "prompts")]
 pub mod prompts;
 mod reporter;
+mod stream;
 #[cfg(feature = "ui")]
 pub mod theme;
 #[cfg(feature = "ui")]
@@ -13,3 +14,4 @@ pub mod ui;
 pub use buffer::*;
 pub use console::*;
 pub use reporter::*;
+pub use stream::*;
